+++ public/components/app.tsx
@@ -22,7 +22,7 @@ import { Home as TraceAnalyticsHome } from './trace_analytics/home';
 import { Home as CustomPanelsHome } from './custom_panels/home';
 import { CustomPanelView } from './custom_panels/custom_panel_view';
 import { Home as EventExplorerHome } from './explorer/home';
-import { LogExplorer } from './explorer/logExplorer';
+import { LogExplorer } from './explorer/log_explorer';
 import { observabilityTitle } from '../../common';
 
 interface ObservabilityAppDeps {
