+++ public/components/explorer/data_grid.scss
