+++ public/components/explorer/hooks/index.ts
@@ -9,5 +9,5 @@
  * GitHub history for details.
  */
 
-export { useFetchEvents } from './useFetchEvents';
-export { useFetchVisualizations } from './useFetchVisualizations';
+export { useFetchEvents } from './use_fetch_events';
+export { useFetchVisualizations } from './use_fetch_visualizations';
