+++ public/components/explorer/log_explorer.scss
