+++ public/components/explorer/reducers/query_reducers.ts
