+++ public/components/explorer/slices/query_result_slice.ts
@@ -13,7 +13,7 @@ import {
   createSlice
 } from '@reduxjs/toolkit';
 import { fetchSuccess as fetchSuccessReducer } from '../reducers'
-import { initialTabId } from '../../../framework/redux/store/sharedState';
+import { initialTabId } from '../../../framework/redux/store/shared_state';
 import { 
   REDUX_EXPL_SLICE_QUERY_RESULT
 } from '../../../../common/constants/explorer';
