+++ public/components/explorer/slices/query_slice.ts
@@ -12,7 +12,7 @@
 import { 
   createSlice
 } from '@reduxjs/toolkit';
-import { initialTabId } from '../../../framework/redux/store/sharedState';
+import { initialTabId } from '../../../framework/redux/store/shared_state';
 import { 
   RAW_QUERY,
   REDUX_EXPL_SLICE_QUERIES
