+++ public/components/explorer/visualizations/count_distribution/index.ts
@@ -9,4 +9,4 @@
  * GitHub history for details.
  */
 
-export * from './countDistribution';
+export * from './count_distribution';
