+++ public/components/explorer/visualizations/frame_layout.scss
