+++ public/components/explorer/visualizations/index.tsx
@@ -14,7 +14,7 @@ import './app.scss';
 import _ from 'lodash';
 
 import React from 'react';
-import { FrameLayout } from './frameLayout';
+import { FrameLayout } from './frame_layout';
 import { DataPanel } from './datapanel';
 import { WorkspacePanel } from './workspace_panel';
 import { ConfigPanelWrapper } from './config_panel';
