+++ public/components/explorer/visualizations/workspace_panel/chart_switch.tsx
