+++ public/framework/redux/store/shared_state.ts
