+++ server/plugin.ts
@@ -20,7 +20,7 @@ import {
 
 import { ObservabilityPluginSetup, ObservabilityPluginStart } from './types';
 import { setupRoutes } from './routes/index';
-import { PPLPlugin } from './adaptors/pplPlugin';
+import { PPLPlugin } from './adaptors/ppl_plugin';
 
 export class ObservabilityPlugin implements Plugin<ObservabilityPluginSetup, ObservabilityPluginStart> {
   private readonly logger: Logger;
